rename: folder component components 폴더로 위치 변경
