+++ components/index.tsx
@@ -11,6 +11,13 @@ import { ShareModal } from "./ShareModal";
 import { AddLinkModal } from "./AddLinkModal";
 import { SearchInput } from "./SearchInput";
 import { Input } from "./Input";
+import { FolderAddLinkArea } from "./FolderAddLinkArea";
+import { FolderCard } from "./FolderCard";
+import { FolderCategory } from "./FolderCategory";
+import { FolderControl } from "./FolderControl";
+import { FolderHeader } from "./FolderHeader";
+import { FolderMain } from "./FolderMain";
+
 export {
   Header,
   CardList,
@@ -25,4 +32,10 @@ export {
   AddLinkModal,
   SearchInput,
   Input,
+  FolderAddLinkArea,
+  FolderCard,
+  FolderCategory,
+  FolderControl,
+  FolderHeader,
+  FolderMain,
 };
