--- pages/folder/components/index.tsx
@@ -1,6 +1,0 @@
-import { FolderAddLinkArea } from "./FolderAddLinkArea";
-import { FolderCategory } from "./FolderCategory";
-import { FolderControl } from "./FolderControl";
-import { FolderHeader } from "./FolderHeader";
-
-export { FolderAddLinkArea, FolderCategory, FolderControl, FolderHeader };
