+++ pages/folder/index.tsx
@@ -1,6 +1,6 @@
 import { Footer } from "@/components";
-import { FolderHeader } from "./components";
-import FolderMain from "./components/FolderMain";
+import { FolderHeader } from "@/components";
+import { FolderMain } from "@/components/FolderMain";
 
 export default function Folder() {
   return (
